moved some files around and in progress of fixing/cleaning up current suite of tests
